fix(apiserver): rename identityproviders to useridentities
Fix typo in resource registration where the endpoint was registered as
"identityproviders" instead of "useridentities", preventing the
UserIdentity resource from being properly exposed in the API.

Changes:
- Rename directory: identityproviders → useridentities
- Update package name from identityproviders to useridentities
- Fix storage map registration to use "useridentities" key
- Update imports in command.go

This aligns the resource name with the actual type name (UserIdentity)
defined in the Milo API schema.

Fixes: Resource not appearing in kubectl api-resources
